Rename domaindns package to dnsclient
The package is a resolver-backed client for ownership checks, so
dnsclient matches its shape better than domaindns.
