+++ pkg/dnsclient/cname.go
@@ -18,7 +18,7 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
-package domaindns
+package dnsclient
 
 import (
 	"context"
